# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /price-list:
    get:
      operationId: listPriceList
      tags: [price-list]
      summary: List price-list items
      responses:
        "200":
          description: All price-list items (active + inactive)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceListItem"
    post:
      operationId: createPriceListItem
      tags: [price-list]
      summary: Add a new price-list item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceListInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceListItem"
        "409":
          description: Duplicate name
  /price-list/{id}:
    patch:
      operationId: updatePriceListItem
      tags: [price-list]
      summary: Edit a price-list item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceListInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceListItem"
        "404":
          description: Not found
    delete:
      operationId: deactivatePriceListItem
      tags: [price-list]
      summary: Soft-delete (deactivate) a price-list item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceListItem"
        "404":
          description: Not found
...
components:
  schemas:
...
    PriceListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        priceCents:
          type: integer
        active:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, priceCents, active, sortOrder, createdAt, updatedAt]
    PriceListInput:
      type: object
      properties:
        name:
          type: string
        priceCents:
          type: integer
        active:
          type: boolean
        sortOrder:
          type: integer
...