# lib/api-zod/src/generated/types/priceListItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceListItem {
  id: number;
  name: string;
  priceCents: number;
  active: boolean;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}